Cleanups
git-svn-id: svn+ssh://solsvn.oma.be/home/sidc/ssr/SWHV_repos/trunk/client/jhv-3d-wcs@7775 b4e469a2-07ce-4b26-9273-4d7d95a670c7

+++ src/jhv/src/org/helioviewer/jhv/base/FileUtils.java
@@ -360,8 +360,8 @@ public class FileUtils {
     }
 
     public static boolean deleteDir(File dir) {
-        if (dir.isDirectory()) {
-            String[] children = dir.list();
+        String[] children;
+        if (dir.isDirectory() && (children = dir.list()) != null) {
             for (String child : children) {
                 boolean success = deleteDir(new File(dir, child));
                 if (!success) {
